+++ nemo_skills/evaluation/evaluate_results.py
@@ -12,21 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import json
 import logging
-import shutil
-import subprocess
 import sys
-from argparse import Namespace
 from dataclasses import field
-from pathlib import Path
 from typing import Any
 
 import hydra
-from omegaconf import MISSING, OmegaConf
+from omegaconf import MISSING
 
-from nemo_skills.code_execution.sandbox import get_sandbox, sandbox_params
-from nemo_skills.evaluation.code_utils import preprocess_code
+from nemo_skills.code_execution.sandbox import sandbox_params
 from nemo_skills.evaluation.settings import GRADING_MAP
 from nemo_skills.utils import get_help_message, nested_dataclass, setup_logging
 
@@ -44,7 +38,7 @@ class EvaluateResultsConfig:
     # Sandbox configuration {sandbox_params}
     sandbox: dict = field(default_factory=lambda: {'sandbox_type': 'local'})
 
-    eval_type: str = "math"  # math or code
+    eval_type: str = "math"  # math or code  TODO: benchmark?
     eval_config: dict = field(default_factory=dict)
 
     def __post_init__(self):
@@ -55,88 +49,6 @@ class EvaluateResultsConfig:
 
 cs = hydra.core.config_store.ConfigStore.instance()
 cs.store(name="base_evaluate_results_config", node=EvaluateResultsConfig)
-
-
-def math_eval(cfg):
-    sandbox = get_sandbox(**cfg.sandbox)
-    sandbox.batch_evaluate_results(
-        prediction_jsonl_files=cfg.prediction_jsonl_files,
-        **cfg.eval_config,
-    )
-
-
-def code_eval(cfg):
-    # TODO: need to move it to a separate docker (either our sandbox or separate srun)
-    from evalplus.evaluate import evaluate
-
-    # processing each generation separately (TODO: evalplus can do it together, but need to figure out the format)
-    for jsonl_file in cfg.prediction_jsonl_files:
-        with open(jsonl_file) as f:
-            samples = [preprocess_code(json.loads(line)) for line in f]
-        # all changes will be done with a new key "completion", so it's ok to write to the same file
-        with open(jsonl_file, "wt", encoding="utf-8") as f:
-            for sample in samples:
-                f.write(json.dumps(sample) + "\n")
-        eval_config = {
-            "samples": jsonl_file,
-            "base_only": False,
-            "parallel": None,
-            "i_just_wanna_run": False,
-            "test_details": False,
-            "min_time_limit": 1,
-            "gt_time_limit_factor": 4.0,
-            "mini": False,
-            "noextreme": False,
-            "version": "default",
-        }
-        eval_config.update(OmegaConf.to_container(cfg.eval_config))
-        evaluate(Namespace(**eval_config))
-        with open(jsonl_file[:-6] + '_eval_results.json', 'rt', encoding="utf-8") as fin:
-            evalplus_grades = json.load(fin)
-        # adding is_correct key to allow compute_metrics to work
-        with open(jsonl_file, "wt", encoding="utf-8") as f:
-            for sample in samples:
-                sample['is_correct'] = evalplus_grades['eval'][sample['task_id']][0]['base_status'] == "pass"
-                sample['is_correct-plus'] = (
-                    sample['is_correct'] and evalplus_grades['eval'][sample['task_id']][0]['plus_status'] == "pass"
-                )
-                f.write(json.dumps(sample) + "\n")
-
-        # moving eval file as otherwise evalplus does not want to recompute metrics if it's present..
-        shutil.move(jsonl_file[:-6] + '_eval_results.json', jsonl_file[:-6] + '_eval_results-saved.json')
-
-
-def ifeval(cfg):
-    for jsonl_file in cfg.prediction_jsonl_files:
-        parent_dir = Path(jsonl_file).absolute().parent
-        cmd = (
-            'cd /opt/benchmarks/google-research && python -m instruction_following_eval.evaluation_main '
-            f'--input_data={jsonl_file} '
-            f'--input_response_data={jsonl_file} '
-            f'--output_dir={parent_dir} '
-        )
-        subprocess.run(cmd, shell=True, check=True)
-        # fusing eval metrics back into the generation file
-        with open(jsonl_file, "rt", encoding="utf-8") as f:
-            samples = [json.loads(line) for line in f]
-
-        with open(parent_dir / 'eval_results_loose.jsonl', 'rt', encoding="utf-8") as f:
-            eval_results = [json.loads(line) for line in f]
-        for sample, eval_result in zip(samples, eval_results):
-            sample['loose_eval'] = eval_result
-
-        with open(parent_dir / 'eval_results_strict.jsonl', 'rt', encoding="utf-8") as f:
-            eval_results = [json.loads(line) for line in f]
-        for sample, eval_result in zip(samples, eval_results):
-            sample['strict_eval'] = eval_result
-
-        with open(jsonl_file, "wt", encoding="utf-8") as f:
-            for sample in samples:
-                f.write(json.dumps(sample) + "\n")
-
-        # removing metric files to avoid reusing them
-        (parent_dir / 'eval_results_loose.jsonl').unlink()
-        (parent_dir / 'eval_results_strict.jsonl').unlink()
 
 
 @hydra.main(version_base=None, config_name="base_evaluate_results_config")
